Phase 2 steps 2/10/11: restructure modules/flash/tune, move FlashKernel
Collapse the old modules/flash/tune/flash/ (kernel-level) and
modules/flash/tune/flash_op/ (op-level) subpackages, split by tuning
level as a module-name suffix, into a single flat modules/flash/tune/
package where 'kernel' vs 'op' is a TuningLevel axis instead
(modular-tune.md §4.3):

  - desc.py / entry.py: shared FlashEntry/FlashInputMetadata description
  - calls.py / reference.py / sdpa_math.py: relocated call wrappers,
    reference implementation, and math helpers (renamed from
    flash/kernel_calls.py, flash/reference.py, flash/utils.py)
  - level_kernel.py / level_op.py: the two TuningLevel implementations
    (list_impls/get_impl/enumerate_variants/impl_desc), replacing
    flash/kernel_opts.py + flash/kernels.py and flash_op/kernels.py +
    flash_op/module.py
  - sancheck.py: FlashKernel.sancheck_lut_tensor/_gen_missing_entries
    moved here from modules/flash/aot/_common.py (step 11) -- verified
    byte-identical output against the old TUNE_V3BIS missing-entries
    print path with a synthetic fully-missing LUT tensor
  - python/template_instantiation/ir/kdesc.py: back-edge updated to the
    new sancheck.py location
