+++ modules/flash/aot/_common.py
@@ -3,10 +3,6 @@
 
 import os
 
-from aotriton.gpu_targets import AOTRITON_ARCH_WARPSIZE
-from aotriton.utils import log
-
-
 # Shared across the flash kernel descriptions (attn_fwd / bwd_kernel_* / bwd_preprocess*).
 MAIN_DTYPES = ['*fp16:16', '*bf16:16', '*fp32:16']
 
@@ -51,117 +47,11 @@ def check_value(functional, repr_name):
             return tc[aname].triton_compile_signature
     assert False, f'Cannot find {repr_name=} in {functional=}'
 
-# TODO: unify FlashKernel with modules/flash/tune/'s FlashEntry/FlashKernelSelector
-#       (Phase 2: modularization unification, see agent-plans/modular-tune.md).
-class FlashKernel:
-    """Flash-family LUT sancheck + missing-entry diagnostic, called by the ATI
-    kdesc (FlashKernel.method(self=kdesc, ...) via family_aot). A plain holder — no
-    description base; it relies only on the duck-typed kdesc surface (check_value,
-    gen_autotune_configs presence)."""
-    FAMILY = 'flash'
-    LUT_FULL_SEQLEN_Q = [16,32,64,128,256,512,1024,2048,4096,8192]
-    LUT_FULL_SEQLEN_K = [16,32,64,128,256,512,1024,2048,4096,8192]
-    LUT_FULL_SEQLEN_NAVI = [16,32,64,128,256,512,1024,2048]
-
-    def is_functional_disabled(self, functional):
-        if not hasattr(self, 'gen_autotune_configs'):  # only check acutal FA kernels
-            return False
-        is_causal = check_value(functional, ['CAUSAL', 'CAUSAL_TYPE'])
-        bias_type = check_value(functional, 'BIAS_TYPE')
-        # print(f'Functional {functional.godel_number=} {is_causal=} {bias_type=}')
-        if is_causal and bias_type != 0:
-            return True
-        if functional.arch.startswith('gfx11'):
-            hdim = check_value(functional, 'BLOCK_DMODEL')
-            if hdim > 256:
-                return True
-        return False
-
-    def sancheck_lut_tensor(self,
-                            functional : 'Functional',
-                            lut_tensor):
-        # Only kernels that provide gen_autotune_configs may have entries in
-        # tuning database
-        if not hasattr(self, 'gen_autotune_configs'):
-            return True, [], _empty_generator()
-        arch = functional.arch
-        if self.is_functional_disabled(functional):
-            return True, [], _empty_generator()
-        MI = (AOTRITON_ARCH_WARPSIZE[arch] == 64)
-        Navi = (AOTRITON_ARCH_WARPSIZE[arch] == 32)
-        LUT_TENSOR_SIZE = (len(self.LUT_FULL_SEQLEN_Q), len(self.LUT_FULL_SEQLEN_K))
-        LUT_TENSOR_SIZE_NAVI = (len(self.LUT_FULL_SEQLEN_NAVI), len(self.LUT_FULL_SEQLEN_NAVI))
-        log(lambda : f'{lut_tensor.shape=} ==? {LUT_TENSOR_SIZE=}')
-        all_pos = (lut_tensor >= 0).all()
-        shape = lut_tensor.shape[1:]
-        if MI:
-            shape_match = shape == LUT_TENSOR_SIZE
-        elif Navi:
-            shape_match = (shape == LUT_TENSOR_SIZE or shape == LUT_TENSOR_SIZE_NAVI)
-        else:
-            assert False, f"Unknown {arch}"
-        ok = all_pos and shape_match
-        if ok:
-            return ok, [], _empty_generator()
-        errors = []
-        if not all_pos:
-            errors.append("certain entries are empty (-1)")
-        if not shape_match:
-            if Navi:
-                errors.append(f"Unexpected {shape=}, Expecting {LUT_TENSOR_SIZE} or {LUT_TENSOR_SIZE_NAVI}")
-            else:
-                errors.append(f"Unexpected {shape=}, Expecting {LUT_TENSOR_SIZE}")
-        # Pick the seqlen lists that match the actual lut_tensor shape for this arch.
-        if Navi and lut_tensor.shape[1:] == LUT_TENSOR_SIZE_NAVI:
-            lut_full_seqlen_q = self.LUT_FULL_SEQLEN_NAVI
-            lut_full_seqlen_k = self.LUT_FULL_SEQLEN_NAVI
-            expected_size = LUT_TENSOR_SIZE_NAVI
-        else:
-            lut_full_seqlen_q = self.LUT_FULL_SEQLEN_Q
-            lut_full_seqlen_k = self.LUT_FULL_SEQLEN_K
-            expected_size = LUT_TENSOR_SIZE
-        missing_entries = self._gen_missing_entries(functional, lut_tensor,
-                                                    arch, lut_full_seqlen_q,
-                                                    lut_full_seqlen_k, expected_size)
-        return ok, errors, missing_entries
-
-    def _gen_missing_entries(self, functional, lut_tensor,
-                             arch, lut_full_seqlen_q, lut_full_seqlen_k, expected_size):
-        import numpy as np
-        from .flash_entry import FlashEntry
-        causal_raw = check_value(functional, 'CAUSAL_TYPE')
-        hdim = check_value(functional, 'BLOCK_DMODEL')
-        dropout_p = 0.5 if check_value(functional, 'ENABLE_DROPOUT') else 0.0
-        q_ptr = check_value(functional, 'Q')
-        if q_ptr.startswith('*fp16'):
-            dtype = 'float16'
-        elif q_ptr.startswith('*bf16'):
-            dtype = 'bfloat16'
-        else:
-            dtype = 'float32'
-        bias_type = check_value(functional, 'BIAS_TYPE')
-        causal = bool(causal_raw)  # 0 → False, non-zero → True
-        def make_entry(seqlen_q, seqlen_k) -> str:
-            entry = FlashEntry(
-                dtype=dtype,
-                hdim=hdim,
-                seqlen_q=seqlen_q,
-                seqlen_k=seqlen_k,
-                causal=causal,
-                dropout_p=dropout_p,
-                bias_type=bias_type,
-            )
-            return f'arch={arch} {entry.as_text()}'
-        if lut_tensor.size == 1:
-            for seqlen_q in lut_full_seqlen_q:
-                for seqlen_k in lut_full_seqlen_k:
-                    yield make_entry(seqlen_q, seqlen_k)
-        else:
-            # TODO: support non-mod0
-            if lut_tensor.shape[1:] == expected_size:
-                _, M_idxs, N_idxs = np.where(lut_tensor < 0)
-            else:
-                fake_lut = np.full(expected_size, -1, dtype=np.int32)
-                M_idxs, N_idxs = np.where(fake_lut < 0)
-            for M_id, N_id in zip(M_idxs, N_idxs):
-                yield make_entry(lut_full_seqlen_q[M_id], lut_full_seqlen_k[N_id])
+# NOTE: FlashKernel (LUT sancheck + missing-entry diagnostic) moved to
+# modules/flash/tune/sancheck.py (modular-tune.md §3b/step 11) -- the codegen
+# back-edge in python/template_instantiation/ir/kdesc.py now resolves it via
+# aotriton.tune.registry.load_family_tune('flash').sancheck.FlashKernel
+# instead of this module. check_value/_empty_generator stay here too (small,
+# pure, stable helpers duplicated rather than imported, same pattern as
+# aot/flash_entry.py vs. tune/entry.py) since other aot/*.py files
+# (aiter_fwd.py, aiter_bwd.py) still depend on check_value from here.
