+++ modules/flash/tune/__init__.py
@@ -1,4 +1,4 @@
-# Copyright © 2026 Advanced Micro Devices, Inc.
+# Copyright © 2025-2026 Advanced Micro Devices, Inc.
 # SPDX-License-Identifier: MIT
 
 """
@@ -10,14 +10,24 @@ python/codegen/parser.py's `load_family_aot` loads modules/<family>/aot/. This
 keeps `modules/<family>` a plain directory rather than a package, so a stray
 `flash.py` on sys.path cannot shadow it.
 
-Phase 1 (pure relocation, no unification): this package holds two independent
-subpackages moved verbatim from v3python.tune -- `flash` (kernel-level tuning)
-and `flash_op` (operator-level tuning) -- each still exporting its own
-`TuneDesc`/`ImplSelector`. Nothing is re-exported here yet; the registry
-resolves 'flash' -> this package's `flash` submodule and 'flash_op' -> its
-`flash_op` submodule.
+Phase 2 (modularization unification, modular-tune.md §4.1-§4.3): the former
+Phase-1 `flash` (kernel-level) and `flash_op` (op-level) subpackages are
+unified into a single `FlashTune` description with a `tuning_level` axis
+('kernel' | 'op'), selected via `FlashTune(level=...)`. `ImplSelector` is the
+single DSL replacing the old per-module FlashKernelSelector/
+FlashOpBackendSelector pair (see aotriton.tune.tdesc).
 
-Phase 2 unifies the two into a single `FlashTune` description (one `TuneDesc`,
-`LEVELS = {'kernel': ..., 'op': ...}`) — see modular-tune.md §4.1/§4.3. This
-file's shape will change then.
+Everything below except `sancheck` (needed eagerly by the codegen back-edge in
+python/template_instantiation/ir/kdesc.py, and torch-free) stays lazily
+resolved: importing this package must not pull in torch/pyaotriton (see
+desc.py's module docstring for why).
 """
+
+from .desc import FlashTune
+from aotriton.tune.tdesc import ImplSelector
+from . import sancheck
+
+TuneDesc = FlashTune
+LEVELS = FlashTune.LEVELS
+
+__all__ = ['TuneDesc', 'ImplSelector', 'LEVELS', 'sancheck']
