--- modules/flash/tune/flash/__init__.py
@@ -1,7 +1,0 @@
-# Copyright © 2025 Advanced Micro Devices, Inc.
-# SPDX-License-Identifier: MIT
-
-from .module import (
-    Flash as TuneDesc,
-    FlashKernelSelector as ImplSelector,
-)
