--- modules/flash/tune/flash_op/__init__.py
@@ -1,7 +1,0 @@
-# Copyright © 2026 Advanced Micro Devices, Inc.
-# SPDX-License-Identifier: MIT
-
-from .module import (
-    FlashOp as TuneDesc,
-    FlashOpBackendSelector as ImplSelector,
-)
